+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Change log
 
 - Support for PHP 7.1.
 - Support for logging, using [monolog/monolog][monolog] by default.
-- Integration test suite.
+- Functional test suite.
 
 ### Removed